Moved migration script to migration folder with upgrade/downgrade methods. Note: For large databases where the graphs count is high like ours, elasticsearch limit of total fields is defaulted to 1000. The new migrations script sents a PUT request to the local running elasticsearch instance to increase the total fields limit to 10000
